Make list of conditions easier to read.
I would like to change this pattern:

```
a == null && b != null,
a != null && b == null,
a != null && b != null && a.equals(b)
```

to simply:
```
!Objects.equals(a,b)
```

The only thing holding me back is that I believe they would behave differently if both a and b are null. In the first version, it would return false, but in the second version, it would return true.

However, this might be a bug? The description in the comments above the method suggest that if both are null, it should return true.