ZFIN-8392: Remove duplicates from MarkerEditDbLinks dialog
Add the fdb_db_name column to the serialization of ForeignDB object. It already pulled that column, but transformed it to an enum.  This adds the column again, but as a string.

Remove duplicate entries from the dialog for MarkerEditDbLink (of identical options, favor the earliest zdbID).

Clean up the DisplayGroupRepository for modern hibernate.

lombokify where appropriate